fixed all memory leaks (YIPEEEEEE), fixed most warnings, change pseudo health & stamina shit

+++ CameraSystem.cpp
@@ -11,7 +11,7 @@ void CameraSystem::Init() {
 	
 }
 
-void CameraSystem::Update(float dt) {
+void CameraSystem::Update(double dt) {
 
 	if (lerpToTarget)
 	{
